LayoutTests/http/tests/security/contentSecurityPolicy/1.1/child-src/frame-allowed.html

<!DOCTYPE html>
<html>
<head>
    <script src="/js-test-resources/js-test-pre.js"></script>
    <script src="/security/contentSecurityPolicy/resources/child-src-test.js"></script>
    <meta http-equiv="Content-Security-Policy" content="child-src 'self';">
</head>
<body>
    <script>
        description("Frames should be governed by 'child-src'.");
        injectFrame("/security/contentSecurityPolicy/resources/alert-pass.html", EXPECT_LOAD);
    </script>
    <script src="/js-test-resources/js-test-post.js"></script>
</body>
</html>
