LayoutTests/http/tests/security/contentSecurityPolicy/1.1/child-src/frame-allowed-expected.txt

CONSOLE MESSAGE: The Content Security Policy directive 'child-src' has been deprecated. Instead, you should use 'worker-src'.
CONSOLE MESSAGE: The Content Security Policy directive 'child-src' has been deprecated. Instead, you should use 'worker-src'.
ALERT: PASS
CONSOLE MESSAGE: IFrame load event fired: the IFrame's location is 'http://127.0.0.1:8000/security/contentSecurityPolicy/resources/alert-pass.html'.
Frames should be governed by 'child-src'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS The IFrame should not have been blocked. It wasn't.
PASS successfullyParsed is true

TEST COMPLETE
